move action framework to toolkit-lib
